refactor: editor -> init -> pipeline ===> editor -> pipeline
