+++ lua/tirenvi/editor/autocmd.lua
@@ -1,9 +1,9 @@
 -- dependencies
 local config = require("tirenvi.config")
+local pipeline = require("tirenvi.app.pipeline")
 local Context = require("tirenvi.app.context")
 local Bufline = require("tirenvi.parser.bufline")
 local buffer = require("tirenvi.io.buffer")
-local init = require("tirenvi.init")
 local buf_state = require("tirenvi.io.buf_state")
 local ui = require("tirenvi.ui")
 local guard = require("tirenvi.util.guard")
@@ -51,7 +51,8 @@ local function on_lines(_, bufnr, tick, range3, bytecount)
 	if buf_state.should_skip(bufnr) then return end
 	local ctx = get_context(bufnr)
 	Debug.ui_entry(bufnr, Range3.short(range3))
-	init.on_lines(ctx, range3)
+	pipeline.on_lines(ctx, range3)
+	pipeline.check_and_repair(ctx, range3)
 	Debug.ui_exit(bufnr, Range3.short(range3))
 end
 
@@ -87,28 +88,28 @@ end
 
 ---@param ctx Context
 local function on_insert_leave(ctx)
-	init.on_insert_leave(ctx)
+	pipeline.check_and_repair(ctx)
 end
 
 ---@param ctx Context
 local function on_buf_read_post(ctx)
 	buffer.clear_buf_local(ctx.bufnr)
-	init.read_post(ctx)
+	pipeline.read_post(ctx)
 end
 
 ---@param ctx Context
 local function on_buf_write_pre(ctx)
-	init.write_pre(ctx)
+	pipeline.write_pre(ctx)
 end
 
 ---@param ctx Context
 local function on_buf_write_post(ctx)
-	init.write_post(ctx)
+	pipeline.write_post(ctx)
 end
 
 ---@param ctx Context
 local function on_insert_char_pre(ctx)
-	init.insert_char_in_newline(ctx)
+	pipeline.insert_char_in_newline(ctx)
 end
 
 ---@param args table
@@ -122,12 +123,12 @@ local function on_cursor_moved(args)
 	if vim.log.levels.DEBUG >= config.log.level then
 		Debug.show_attr_marks(ctx)
 	end
-	init.auto_wrap(ctx)
+	pipeline.auto_wrap(ctx)
 end
 
 ---@param ctx Context
 local function on_filetype(ctx)
-	init.on_filetype(ctx)
+	pipeline.on_filetype(ctx)
 end
 
 ---@param args table
